feat: Add mockFetch
